fix tests

+++ cmd/createCollection_test.go
@@ -134,7 +134,7 @@ func TestExecuteCreateCollection(t *testing.T) {
 		aggregationErr       error
 		power                int8
 		powerErr             error
-		tolerance            uint16
+		tolerance            uint32
 		toleranceErr         error
 		createCollectionErr  error
 		createCollectionHash common.Hash
